% Generated by roxygen2 (4.0.0): do not edit by hand
\name{api_key}
\alias{api_key}
\title{Retrieve the difftron api key.}
\usage{
api_key()
}
\description{
If not supplied explicitly, this should be stored in an environment
variable called \code{DIFFTRON_KEY}. Set this in you \code{.Renviron}.
If you set in a project specific {.Renviron}, make sure you exclude from
git etc.
}

