% Generated by roxygen2 (4.0.0): do not edit by hand
\name{put_imageset}
\alias{put_imageset}
\title{PUT an imageset.}
\usage{
put_imageset(project, set, path, key = api_key())
}
\arguments{
  \item{project,set}{Name of project and image set}

  \item{path}{character vector of images to upload}

  \item{key}{difftron api key, see \code{\link{api_key}}
  for more details}
}
\description{
PUT an imageset.
}
\examples{
\donttest{
png("test.png"); plot(runif(10)); dev.off()
put_imageset("test", "test2", "test.png")
unlink("test.png")
}
}
\keyword{internal}

