% Generated by roxygen2 (4.0.0): do not edit by hand
\name{riff_dir}
\alias{riff_dir}
\alias{riff_package}
\title{Run all R files in a directory or package and upload visual results.}
\usage{
riff_dir(path, env = NULL, ...)

riff_package(package, ...)
}
\arguments{
  \item{path}{Directory of R files to execute}

  \item{package}{Name of package. All files in
  \cpde{tests/rifftron} will be executed, in the package
  environment (so that non-exported functions can be
  tested).}

  \item{env}{Environment in which to execute code. If
  \code{NULL}, defaults to a new environment inheriting
  from the global environment.}

  \item{...}{Additional arguments passed on to \code{}}
}
\description{
Run all R files in a directory or package and upload visual results.
}
\examples{
\dontest{
riff_package("rifftron")
}
}

