% Generated by roxygen2 (4.0.0): do not edit by hand
\name{set_diff_dir}
\alias{get_diff_dir}
\alias{set_diff_dir}
\title{Get and set diff directory.}
\usage{
set_diff_dir(path, create = FALSE)

get_diff_dir()
}
\arguments{
  \item{path}{path in which to store temporary images}

  \item{create}{If \code{TRUE} will create directory if
  needed, otherwise will throw error if directory is
  missing}
}
\value{
\code{set_diff_dir()} invisibly returns old path.
}
\description{
By default, all images created by rifftron will be stored in this
directory, so you can easily upload a complete image set in a single
step.  This defaults to a session-specific temporary directory.
}

