% Generated by roxygen2 (4.0.0): do not edit by hand
\name{sha1}
\alias{sha1}
\title{Determine sha1 of the current git checkout}
\usage{
sha1(n = 10)
}
\arguments{
  \item{n}{Number of characters to truncate sha1 to.
  Defaults to 10 because that's what github does.}
}
\description{
A \code{-uncommited} suffix is added if the repo currently has uncommited
changes that are not reflected in the SHA.
}

