% Generated by roxygen2 (4.0.0): do not edit by hand
\name{upload_dir}
\alias{upload_dir}
\title{Upload all files in a directory.}
\usage{
upload_dir(set_name = sha1(), project = project_name(),
  path = get_diff_dir(), key = api_key())
}
\arguments{
  \item{set_name}{Name of the image set. Defaults to the
  \code{\link{sha1}} of the current git commit. Manually
  override if you're not using git.}

  \item{project}{Name of the project. Defaults to the name
  of the current directory}

  \item{path}{Path where images are located. All files
  ending in \code{.png} are uploaded. Defaults to
  \code{\link{get_diff_dir()}}}

  \item{key}{difftron api get. See \code{\link{api_key}}
  for more details.}
}
\description{
Upload all files in a directory.
}
\examples{
\donttest{
set_diff_dir(tempfile(), create = TRUE)
capture_plot("Sequence", plot(1:10))
capture_plot("Random", plot(runif(10)))
upload_dir()
}
}

