% Generated by roxygen2 (4.0.0): do not edit by hand
\name{capture_plot}
\alias{capture_plot}
\title{Run some code, capturing an image.}
\usage{
capture_plot(code, path, dev = "png", width = 250, height = 250, ...)
}
\arguments{
  \item{code}{to evaluate and print. Printing ensure that
  it works easily with ggplot2 and lattice graphics.}

  \item{path}{where to save resulting image}

  \item{dev}{name of graphics device to use, defaults to
  \code{\link{png}}.}

  \item{width,height}{dimenion, in pixels, of resulting
  plot}

  \item{...}{other arguments passed on to the graphics
  device.}
}
\value{
(invisibly) the path of the created file
}
\description{
Run some code, capturing an image.
}
\examples{
capture_plot(plot(1:10), tempfile(fileext = ".png"))
}

