% Generated by roxygen2 (4.0.0): do not edit by hand
\name{riff_dir}
\alias{riff_dir}
\alias{riff_travis}
\title{Run all R files in a directory or package and upload visual results.}
\usage{
riff_dir(path, env = NULL, ...)

riff_travis()
}
\arguments{
  \item{path}{Directory of R files to execute}

  \item{env}{Environment in which to execute code. If
  \code{NULL}, defaults to a new environment inheriting
  from the global environment.}

  \item{...}{Additional arguments passed on to \code{}}
}
\description{
\code{riff_travis} runs code in the package environment, so that
non-exported functions can be tested.
}
\examples{
\donttest{
riff_travis()
riff_dir("tests/rifftron")
}
}

